Move otspec bit of head table to otspec/tables
