+++ crates/otspec/src/tables.rs
@@ -1,0 +1,2 @@
+/// The `head` (Header) table
+pub mod head;
